# infrastructure/swagger/swagger.yaml
definitions:
  domain.Account:
    properties:
      cpf:
        type: string
      created_at:
        type: string
      id:
        type: string
      string:
        type: string
    type: object
  domain.AccountBalance:
    properties:
      balance:
        type: number
    type: object
  domain.AccountCreatorRequest:
    properties:
      cpf:
        type: string
      name:
        type: string
      secret:
        type: string
    type: object
  domain.AccountCreatorResponse:
    properties:
      id:
        type: string
    type: object
  domain.AccountList:
    properties:
      accounts:
        items:
          $ref: '#/definitions/domain.Account'
        type: array
    type: object
  domain.JwtToken:
    properties:
      token:
        type: string
    type: object
  domain.Login:
    properties:
      cpf:
        type: string
      secret:
        type: string
    type: object
  domain.ResponseError:
    properties:
      message:
        type: string
    type: object
  domain.TransferCreatorResponse:
    properties:
      id:
        type: string
      new_account_destination_balance:
        type: number
      new_account_origin_balance:
        type: number
      old_account_destination_balance:
        type: number
      old_account_origin_balance:
        type: number
    type: object
  domain.TransferRequest:
    properties:
      account_destination_id:
        type: string
      amount:
        type: number
    type: object
info:
  contact: {}
paths:
  /accounts:
    get:
      description: Return the list of all accounts. Fields 'secret' and 'balance'
        are omitted
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AccountList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ResponseError'
      summary: Get accounts
      tags:
      - Accounts
    post:
      consumes:
      - application/json
      description: Create new Account in case 'cpf' doesn't exists yet
      parameters:
      - description: Account request
        in: body
        name: AccountRequest
        required: true
        schema:
          $ref: '#/definitions/domain.AccountCreatorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.AccountCreatorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ResponseError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/domain.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ResponseError'
      summary: Create account
      tags:
      - Accounts
  /accounts/{account_id}/balance:
    get:
      description: Return account balance
      parameters:
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AccountBalance'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ResponseError'
      summary: Get account balance
      tags:
      - Accounts
  /login:
    post:
      consumes:
      - application/json
      description: Authemticate user
      parameters:
      - description: Login request
        in: body
        name: Login
        required: true
        schema:
          $ref: '#/definitions/domain.Login'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.JwtToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ResponseError'
      summary: Login
      tags:
      - Login
  /transfers:
    post:
      consumes:
      - application/json
      description: Transfer amount from origin account to destination account
      parameters:
      - description: Transfer request
        in: body
        name: TransferRequest
        required: true
        schema:
          $ref: '#/definitions/domain.TransferRequest'
      - description: Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.TransferCreatorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ResponseError'
      summary: Transfer amount
      tags:
      - Transfers
swagger: "2.0"